deploy_package_v3/requirements.txt
Django==4.2.7
django-froala-editor
whitenoise
django-taggit
